requirements.txt
gps==3.19
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
pyserial==3.5
setuptools==80.9.0
wheel==0.45.1
ultralytics
